requirements.txt
numpy
colormath
pillow
kmeans
openpyxl